functioning umd build with demo.html
