+++ package.json
@@ -43,11 +43,13 @@
         "@babel/preset-env": "7.26.0",
         "@babel/preset-react": "7.23.3",
         "@babel/preset-typescript": "7.23.3",
+        "@rollup/plugin-alias": "5.1.1",
         "@rollup/plugin-babel": "6.0.4",
         "@rollup/plugin-commonjs": "25.0.7",
         "@rollup/plugin-image": "3.0.3",
         "@rollup/plugin-json": "6.1.0",
         "@rollup/plugin-node-resolve": "15.2.3",
+        "@rollup/plugin-replace": "6.0.2",
         "@rollup/plugin-terser": "0.4.4",
         "@rollup/plugin-typescript": "11.1.6",
         "@rollup/plugin-url": "8.0.2",
