+++ rollup.config.mjs
@@ -1,9 +1,12 @@
+import alias from '@rollup/plugin-alias';
 import babel from '@rollup/plugin-babel';
 import commonjs from '@rollup/plugin-commonjs';
 import dts from 'rollup-plugin-dts';
 import image from '@rollup/plugin-image';
 import json from '@rollup/plugin-json';
 import peerDepsExternal from 'rollup-plugin-peer-deps-external';
+import replace from '@rollup/plugin-replace';
+import path from 'path';
 import postcss from 'rollup-plugin-postcss';
 import resolve from '@rollup/plugin-node-resolve';
 import terser from '@rollup/plugin-terser';
@@ -35,26 +38,11 @@ export default [
                 file: packageJson.main,
                 format: 'cjs',
                 sourcemap: true,
-                exports: 'named',
             },
             {
                 file: packageJson.module,
                 format: 'esm',
                 sourcemap: true,
-                exports: 'named',
-            },
-            {
-                file: 'build/umd/rcpch-digital-growth-charts.umd.js',
-                format: 'umd',
-                name: 'RCPCHGrowthCharts',
-                sourcemap: true,
-                globals: {
-                    react: 'React',
-                    'react-dom': 'ReactDOM',
-                    'react-dom/client': 'ReactDOM',
-                    'styled-components': 'styled',
-                },
-                exports: 'named', // Ensure named exports
             },
         ],
         plugins: [
@@ -68,17 +56,15 @@ export default [
                 ignoreGlobal: true,
                 include: /\/node_modules\//,
             }),
-            babel({
-                exclude: 'node_modules/**',
-                babelHelpers: 'bundled',
-                presets: ['@babel/preset-react', '@babel/preset-typescript'],
-                extensions: ['.ts', '.tsx'],
+            typescript(),
+            terser({
+                compress: {
+                    pure_getters: true, // assume obj.prop has no side effects
+                    dead_code: true,
+                    toplevel: true,
+                },
+                mangle: true,
             }),
-            typescript({
-                // Remove jsx: 'react' here
-                target: 'es5',
-            }),
-            terser(),
             json(),
             versionInjector(),
             image(),
@@ -86,8 +72,62 @@ export default [
     },
     {
         input: 'src/index.ts',
-        output: [{ file: 'build/types.d.ts', format: 'es' }],
         external: [],
+        output: [{ file: 'build/types.d.ts', format: 'es' }],
         plugins: [dts.default()],
+    },
+    {
+        input: 'src/umd.ts',
+        external: ['react', 'react-dom', 'react-dom/client'],
+        output: [
+            {
+                file: 'build/umd/rcpch-digital-growth-charts.umd.js',
+                format: 'umd',
+                name: 'RCPCHGrowthCharts',
+                exports: 'default',
+                globals: {
+                    react: 'React',
+                    'react-dom': 'ReactDOM',
+                    'react-dom/client': 'ReactDOM',
+                },
+                sourcemap: true,
+            },
+        ],
+        plugins: [
+            alias({
+                entries: [
+                    {
+                        find: 'process',
+                        replacement: path.resolve(__dirname, 'src/stubs/process.js'),
+                    },
+                ],
+            }),
+            replace({
+                preventAssignment: true,
+                values: {
+                    'process.env.NODE_ENV': JSON.stringify('production'),
+                },
+            }),
+            postcss({
+                extensions: ['.css'],
+                inject: true,
+            }),
+            resolve(),
+            commonjs({
+                ignoreGlobal: true,
+                include: /\/node_modules\//,
+            }),
+            babel({
+                exclude: 'node_modules/**',
+                babelHelpers: 'bundled',
+                presets: ['@babel/preset-react', '@babel/preset-typescript'],
+                extensions: ['.ts', '.tsx'],
+            }),
+            typescript(),
+            terser(),
+            json(),
+            versionInjector(),
+            image(),
+        ],
     },
 ];
