+++ src/stubs/process.js
@@ -1,0 +1,7 @@
+// This file is used to stub the process object in a Node.js environment.
+// When using Rollup build for UMD, the process object is not available by default so we need to replace it with a stub.
+// There are a bunch of libraries in node_modules that use process.*** so we stub all them here.
+
+export default {
+    env: {},
+};
